renamed db connection script
