+++ backend/api/routes/get_routes/get_wholedata.py
@@ -5,7 +5,7 @@ from fastapi import APIRouter
 
 from app_utils import data_loading
 from app_utils.flooding import add_flood_color
-from query.processed_db import DB
+from query.production_db import DB
 
 logger = logging.getLogger(__name__)
