+++ backend/api/routes/post_routes/post_qcew.py
@@ -3,7 +3,7 @@ from fastapi import APIRouter
 
 from api.metadata_registry import get_metadata
 from api.models import FilterRequest, make_response
-from query.processed_db import DB
+from query.production_db import DB
 
 router = APIRouter()
