+++ backend/app_utils/timeseries_db.py
@@ -6,7 +6,7 @@ Use query_timeseries(table_name, filters) to filter any registered table.
 
 import logging
 
-from query.processed_db import DB
+from query.production_db import DB
 
 logger = logging.getLogger(__name__)
