+++ backend/query/__init__.py
@@ -2,7 +2,7 @@
 
 from query.cdc import dual_var_comparison, get_cdc_county_pca, single_var_geojson
 from query.core_functions import filter_options, filter_tree
-from query.processed_db import DB
+from query.production_db import DB
 from query.wastewater import (
     get_soil_suit_geojson,
     get_waste_service_areas_geojson,
