+++ backend/query/acs5.py
@@ -15,7 +15,7 @@ import pandas as pd
 from api.models import FilterSource, RangeFilter
 from app_utils.sql_render import sql_filter_block
 from query.core_functions import filter_tree
-from query.processed_db import DB
+from query.production_db import DB
 
 logger = logging.getLogger(__name__)
 sql_path = Path(__file__).resolve().parent / "sql" / "acs5"
