+++ backend/query/zoning.py
@@ -16,7 +16,7 @@ import pandas as pd
 
 from api.models import FilterSource
 from app_utils.sql_render import render_sql, sql_filter_block
-from query.processed_db import DB
+from query.production_db import DB
 
 logger = logging.getLogger(__name__)
 sql_dir = Path(__file__).resolve().parent / "sql" / "zoning"
